Provide a way to intercept file system access
As noted, this implementation is imperfect because it’s still based on
std::path, so that it exposes the style of the target OS’s file system
(e.g. C:\ versus /), which will not always be desirable (though it’s a
sound default).

The implementation is also necessarily synchronous, but that’s not a
regression, so I’m not worried about it, even if asynchronous support
could be desirable at some point.

A few situations where this may be useful:

- Security: if you’re handling untrusted Sass source, you should
  definitely restrict file system access, either to limit where it can
  access, or to forbid it altogether, in order to avoid exfiltration.
  (This is certainly necessary for security, but may not be sufficient.)

- Platforms that don’t have a real file system, like wasm, but may have
  a fake file system to facilitate development. (I hope that this will
  gently nudge things in the direction of exposing a uniform API
  regardless of target platform.)

- Operating on any non-materialised file tree, e.g. a remote computer
  via SFTP, or a Git repository without a working tree.

- Doing incremental recompilation, so that you have already read the
  contents of the file and not only don’t *need* to reread it, but
  actively don’t *want* the latest changes if any of the files happen to
  have changed on disk.

(The last one is my situation. Building a static site generator that
uses Salsa for incremental recompilation.)