Optimise import path candidate finding
No need to construct all of the paths up front; better to just test each
one in turn.

I deem the macro warranted; there’s no other way of shortening it (?
means the opposite—return if None, where we want return if Some), and
it’s a clear and simple abstraction.