Don't buffer serializer output
A `Serializer` now receives an instance of `std::io::Write`. This means
that we can no longer write the UTF-8 BOM nor the `@charset "UTF-8"
directive when serialization is done and instead do the ASCII-content
check while visiting the AST.

All `inspect_*` methods now create a temporary buffer and pass it to a
`Serializer`. Then they (unsafely) interpret the buffer as a UTF-8
string. The goal is to avoid this extra set of allocations and always
write directly to the writer passed by the API user.

We also removed some extra allocations marked with todo comments.