Move main file to new `Serializer` API
This also adds a public API to write the output of a compiled set of
SASS files into a `Write`. We should add a recommendation to the
rustdoc about passing a buffered writer; otherwise performance will
probably be horrible.

The previous `String`-based APIs were also moved to this new set of APIs
by creating a temporary `Vec<u8>` buffer, passing it to the serializer,
and finally (unsafely) reinterpreting the buffer contents as a UTF-8
string.